Add functions for determining game version

+++ SC4Fix/dllmain.h
@@ -22,4 +22,4 @@
 #define WIN32_LEAN_AND_MEAN
 #include <Windows.h>
 
-int GetGameVersion(void);
+uint16_t GetGameVersion(void);
